Mongoose Population

+++ routes/dishRouter.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const bodyParser = require('body-parser');
 const mongoose = require('mongoose');
+var authenticate = require('../authenticate');
 
 const Dishes = require('../models/dishes');
 
@@ -9,42 +10,49 @@ const dishRouter = express.Router();
 dishRouter.use(bodyParser.json());
 
 dishRouter.route('/')
-.get((req,res,next) => {
-    Dishes.find({})
-    .then((dishes) => {
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(dishes);
-    }, (err) => next(err))
-    .catch((err) => next(err));
-})
-.post((req, res, next) => {
-    Dishes.create(req.body)
-    .then((dish) => {
-        console.log('Dish Created ', dish);
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(dish);
-    }, (err) => next(err))
-    .catch((err) => next(err));
-})
-.put((req, res, next) => {
-    res.statusCode = 403;
-    res.end('PUT operation not supported on the damn /dishes');
-})
-.delete((req, res, next) => {
-    Dishes.remove({})
-    .then((resp) => {
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(resp);
-    }, (err) => next(err))
-    .catch((err) => next(err));    
-});
+    .get((req,res,next) => {
+        Dishes.find({})
+        .populate('comments.author')
+        .then((dishes) => {
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(dishes);
+        }, (err) => next(err))
+        .catch((err) => next(err));
+    })
+
+    .post(authenticate.verifyUser, (req, res, next) => {
+        authenticate.verifyAdmin(req,res,next);
+        Dishes.create(req.body)
+        .then((dish) => {
+            console.log('Dish Created ', dish);
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(dish);
+        }, (err) => next(err))
+        .catch((err) => next(err));
+    })
+
+    .put(authenticate.verifyUser, (req, res, next) => {
+        res.statusCode = 403;
+        res.end('PUT operation not supported on the damn /dishes');
+    })
+
+    .delete(authenticate.verifyUser, (req, res, next) => {
+        authenticate.verifyAdmin(req,res,next);
+        Dishes.remove({})
+        .then((resp) => {
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(resp);
+        }, (err) => next(err))
+        .catch((err) => next(err));    
+    });
 
 dishRouter.route('/:dishId')
     .get((req,res,next) => {
         Dishes.findById(req.params.dishId)
+        .populate('comments.author')
         .then((dish) => {
             res.statusCode = 200;
             res.setHeader('Content-Type', 'application/json');
@@ -52,11 +60,12 @@ dishRouter.route('/:dishId')
         }, (err) => next(err))
         .catch((err) => next(err));
     })
-    .post((req, res, next) => {
+    .post(authenticate.verifyUser, (req, res, next) => {
         res.statusCode = 403;
         res.end('POST operation not supported on on the damn /dishes/'+ req.params.dishId);
     })
-    .put((req, res, next) => {
+    .put(authenticate.verifyUser, (req, res, next) => {
+        authenticate.verifyAdmin(req,res,next);
         Dishes.findByIdAndUpdate(req.params.dishId, {
             $set: req.body
         }, { new: true })
@@ -67,7 +76,8 @@ dishRouter.route('/:dishId')
         }, (err) => next(err))
         .catch((err) => next(err));
     })
-    .delete((req, res, next) => {
+    .delete(authenticate.verifyUser, (req, res, next) => {
+        authenticate.verifyAdmin(req,res,next);
         Dishes.findByIdAndRemove(req.params.dishId)
         .then((resp) => {
             res.statusCode = 200;
@@ -76,9 +86,11 @@ dishRouter.route('/:dishId')
         }, (err) => next(err))
         .catch((err) => next(err));
     });
+    
 dishRouter.route('/:dishId/comments')
     .get((req,res,next) => {
         Dishes.findById(req.params.dishId)
+        .populate('comments.author')
         .then((dish) => {
             if (dish != null) {
                 res.statusCode = 200;
@@ -93,16 +105,22 @@ dishRouter.route('/:dishId/comments')
         }, (err) => next(err))
         .catch((err) => next(err));
     })
-    .post((req, res, next) => {
+    .post(authenticate.verifyUser, (req, res, next) => {
+        authenticate.verifyAdmin(req,res,next);
         Dishes.findById(req.params.dishId)
         .then((dish) => {
             if (dish != null) {
+                req.body.author = req.user._id;
                 dish.comments.push(req.body);
                 dish.save()
                 .then((dish) => {
-                    res.statusCode = 200;
-                    res.setHeader('Content-Type', 'application/json');
-                    res.json(dish);                
+                    Dishes.findById(dish._id)
+                    .populate('comments.author')
+                    .then((dish) => {
+                        res.statusCode = 200;
+                        res.setHeader('Content-Type', 'application/json');
+                        res.json(dish);
+                    })            
                 }, (err) => next(err));
             }
             else {
@@ -113,12 +131,13 @@ dishRouter.route('/:dishId/comments')
         }, (err) => next(err))
         .catch((err) => next(err));
     })
-    .put((req, res, next) => {
+    .put(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
         res.statusCode = 403;
         res.end('PUT operation not supported on /dishes/'
             + req.params.dishId + '/comments');
     })
-    .delete((req, res, next) => {
+    .delete(authenticate.verifyUser, (req, res, next) => {
+        authenticate.verifyAdmin(req,res,next);
         Dishes.findById(req.params.dishId)
         .then((dish) => {
             if (dish != null) {
@@ -144,6 +163,7 @@ dishRouter.route('/:dishId/comments')
 dishRouter.route('/:dishId/comments/:commentId')
     .get((req,res,next) => {
         Dishes.findById(req.params.dishId)
+        .populate('comments.author')    
         .then((dish) => {
             if (dish != null && dish.comments.id(req.params.commentId) != null) {
                 res.statusCode = 200;
@@ -163,12 +183,13 @@ dishRouter.route('/:dishId/comments/:commentId')
         }, (err) => next(err))
         .catch((err) => next(err));
     })
-    .post((req, res, next) => {
+    .post(authenticate.verifyUser, (req, res, next) => {
         res.statusCode = 403;
         res.end('POST operation not supported on /dishes/'+ req.params.dishId
             + '/comments/' + req.params.commentId);
     })
-    .put((req, res, next) => {
+    .put(authenticate.verifyUser, (req, res, next) => {
+        authenticate.verifyAdmin(req,res,next);
         Dishes.findById(req.params.dishId)
         .then((dish) => {
             if (dish != null && dish.comments.id(req.params.commentId) != null) {
@@ -180,9 +201,13 @@ dishRouter.route('/:dishId/comments/:commentId')
                 }
                 dish.save()
                 .then((dish) => {
-                    res.statusCode = 200;
-                    res.setHeader('Content-Type', 'application/json');
-                    res.json(dish);                
+                    Dishes.findById(dish._id)
+                    .populate('comments.author')
+                    .then((dish) => {
+                        res.statusCode = 200;
+                        res.setHeader('Content-Type', 'application/json');
+                        res.json(dish);  
+                    })              
                 }, (err) => next(err));
             }
             else if (dish == null) {
@@ -198,17 +223,29 @@ dishRouter.route('/:dishId/comments/:commentId')
         }, (err) => next(err))
         .catch((err) => next(err));
     })
-    .delete((req, res, next) => {
+    .delete(authenticate.verifyUser, (req, res, next) => {
         Dishes.findById(req.params.dishId)
         .then((dish) => {
             if (dish != null && dish.comments.id(req.params.commentId) != null) {
-                dish.comments.id(req.params.commentId).remove();
-                dish.save()
-                .then((dish) => {
-                    res.statusCode = 200;
-                    res.setHeader('Content-Type', 'application/json');
-                    res.json(dish);                
-                }, (err) => next(err));
+                var theComment = dish.comments.id(req.params.commentId);
+                if (theComment.author._id === req.user._id){
+                    dish.comments.id(req.params.commentId).remove();
+                    dish.save()
+                    .then((dish) => {
+                        Dishes.findById(dish._id)
+                        .populate('comments.author')
+                        .then((dish) => {
+                            res.statusCode = 200;
+                            res.setHeader('Content-Type', 'application/json');
+                            res.json(dish);  
+                        })               
+                    }, (err) => next(err));
+                }
+                else{
+                    var error = new Error("You are not authorized to delete this comment");
+                    error.status = 403;
+                    return next(error);
+                }
             }
             else if (dish == null) {
                 err = new Error('Dish ' + req.params.dishId + ' not found');
