+++ routes/leaderRouter.js
@@ -5,6 +5,7 @@ const bodyParser = require('body-parser');
 const leaderRouter = express.Router();
 const mongoose = require('mongoose');
 const Leader = require('../models/leaders');
+var authenticate = require('../authenticate');
 
 leaderRouter.use(bodyParser.json());
 
@@ -18,7 +19,8 @@ leaderRouter.route('/')
         }, (err) => next(err))
         .catch((err) => next(err));
     })
-    .post((req, res, next) => {
+    .post(authenticate.verifyUser, (req, res, next) => {
+        authenticate.verifyAdmin(req,res,next);
         Leader.create(req.body)
         .then((leader) => {
             console.log('Leader Created ', promo);
@@ -28,11 +30,12 @@ leaderRouter.route('/')
         }, (err) => next(err))
         .catch((err) => next(err));
     })
-    .put((req, res, next) => {
+    .put(authenticate.verifyUser, (req, res, next) => {
         res.statusCode = 403;
         res.end('PUT operation not supported on /leaders');
     })
-    .delete((req, res, next) => {
+    .delete(authenticate.verifyUser, (req, res, next) => {
+        authenticate.verifyAdmin(req,res,next);
         Leader.remove({})
         .then((removed) => {
             res.statusCode = 200;
@@ -54,11 +57,12 @@ leaderRouter.route('/:leaderId')
         }, (err) => next(err))
         .catch((err) => next(err));
     })
-    .post((req, res, next) => {
+    .post(authenticate.verifyUser, (req, res, next) => {
         res.statusCode = 403;
         res.end('POST operation not supported on /leaders/'+ req.params.leaderId);
     })
-    .put((req, res, next) => {
+    .put(authenticate.verifyUser, (req, res, next) => {
+        authenticate.verifyAdmin(req,res,next);
         Leader.findByIdAndUpdate(req.params.leaderId, {
             $set: req.body
         }, { new: true })
@@ -69,7 +73,8 @@ leaderRouter.route('/:leaderId')
         }, (err) => next(err))
         .catch((err) => next(err));
     })
-    .delete((req, res, next) => {
+    .delete(authenticate.verifyUser, (req, res, next) => {
+        authenticate.verifyAdmin(req,res,next);
         Leader.findByIdAndRemove(req.params.leaderId)
         .then((resp) => {
             res.statusCode = 200;
