+++ routes/promoRouter.js
@@ -4,6 +4,7 @@ const bodyParser = require('body-parser');
 const promoRouter = express.Router();
 const mongoose = require('mongoose');
 const Promo = require('../models/promos');
+var authenticate = require('../authenticate');
 
 promoRouter.use(bodyParser.json());
 
@@ -17,7 +18,8 @@ promoRouter.route('/')
         }, (err) => next(err))
         .catch((err) => next(err));
     })
-    .post((req, res, next) => {
+    .post(authenticate.verifyUser, (req, res, next) => {
+        authenticate.verifyAdmin(req,res,next);
         Promo.create(req.body)
         .then((promo) => {
             console.log('Promo Created ', promo);
@@ -27,11 +29,12 @@ promoRouter.route('/')
         }, (err) => next(err))
         .catch((err) => next(err));
     })
-    .put((req, res, next) => {
+    .put(authenticate.verifyUser, (req, res, next) => {
         res.statusCode = 403;
         res.end('PUT operation not supported on /promos');
     })
-    .delete((req, res, next) => {
+    .delete(authenticate.verifyUser, (req, res, next) => {
+        authenticate.verifyAdmin(req,res,next);
         Promo.remove({})
         .then((removed) => {
             res.statusCode = 200;
@@ -51,11 +54,12 @@ promoRouter.route('/:promoId')
         }, (err) => next(err))
         .catch((err) => next(err));
     })
-    .post((req, res, next) => {
+    .post(authenticate.verifyUser, (req, res, next) => {
         res.statusCode = 403;
         res.end('POST operation not supported on /promos/'+ req.params.dishId);
     })
-    .put((req, res, next) => {
+    .put(authenticate.verifyUser, (req, res, next) => {
+        authenticate.verifyAdmin(req,res,next);
         Dishes.findByIdAndUpdate(req.params.promoId, {
             $set: req.body
         }, { new: true })
@@ -66,7 +70,8 @@ promoRouter.route('/:promoId')
         }, (err) => next(err))
         .catch((err) => next(err));
     })
-    .delete((req, res, next) => {
+    .delete(authenticate.verifyUser, (req, res, next) => {
+        authenticate.verifyAdmin(req,res,next);
         Dishes.findByIdAndRemove(req.params.promoId)
         .then((resp) => {
             res.statusCode = 200;
